C API: simplify CrgContactPointStruct
CrgContactPointStruct contained several members that were never used, or
only used inside a function call. Removing those (and replacing the local
uses with local variables) makes each contact point 96 bytes smaller
(on 64bit systems).

As a nice side effect, this allows sharing a contact point between several
threads, as long as history is disabled. This is important when OpenCRG
is used as part of OpenDRIVE. Consider an OpenDRIVE road which references
several (maybe hundreds) of Object-CRGs. If a thread queries the height of
an OpenDRIVE road, the OpenDRIVE library must query the OpenCRG road object
if it is at the relevant position. Previously, querying the road height
from several threads at the same time was practically impossible using
the OpenCRG C API, since each thread needed a separate contact point per
OpenCRG file, but there is no thread-safe way to acquire a new contact
point at runtime. So the only possible way would have been for the
OpenDRIVE library to acquire as many OpenCRG contact points for each
Object-CRG as simultaneous threads may access the library, and do so at
the very beginning, before any threads access the OpenCRG files.

With this change, OpenDRIVE libraries can safely just store one contact
point for each OpenCRG file, and use that across threads, as long as they
disable history for the thread (which is not a problem for typical
OpenDRIVE use-cases).

With history enabled, behavior is unchanged, but the code is slightly more
efficient than before.