Restore belgian_block.crg file
The faulty \crg-bin\belgian_block.crg could not be read.
It is identical in structure to the file
\c-api\demo\bin\belgian_block.crg
except for the ct block at the beginning of the file.
Comparing the two files shows that the faulty file is missing 6 Bytes within the z-channel data.
This commit repairs the corrupt crg data.

Resolves: #7

Signed-off-by: Patrick Kuemmerle patrick.kuemmerle@3d-mapping.de